+++ src/game_events/event_display.rs
@@ -1,6 +1,5 @@
 use crate::{
-    game_events::{GameEvent, GameEventsState},
-    game_state::RunningState,
+    game_events::{AddEventToCurrentEvent, GameEvent, GameEventsState},
     ui::large_button,
 };
 use bevy::prelude::*;
@@ -18,10 +17,10 @@ pub fn event_display(
     mut egui_context: ResMut<EguiContext>,
     events: Option<Res<HashMap<String, GameEvent>>>,
     mut state: ResMut<GameEventsState>,
-    mut running_state: ResMut<State<RunningState>>,
     mut textures: Local<EventTextures>,
     windows: Res<Windows>,
     asset_server: Res<AssetServer>,
+    mut add_event: EventWriter<AddEventToCurrentEvent>,
 ) {
     if events.is_none() {
         return;
@@ -30,11 +29,7 @@ pub fn event_display(
     let events = events.unwrap();
     let window = windows.primary();
 
-    if state.current_events.is_empty() {
-        running_state.set(RunningState::Running).unwrap();
-    }
-
-    for id in state.current_events.clone().iter() {
+    if let Some(id) = state.current_events.first() {
         let event: &GameEvent = events.get(id).unwrap();
 
         if !textures.textures.contains_key(&event.image) {
@@ -66,9 +61,10 @@ pub fn event_display(
                         for action in &event.actions {
                             let w = ui.available_width();
                             if large_button(ui, w, &action.label).clicked() {
-                                state.current_events.remove(id);
-                                if let Some(id) = &action.trigger_event {
-                                    state.current_events.insert(id.clone());
+                                state.current_events.remove(0);
+                                if let Some(id) = &action.trigger_event.clone() {
+                                    add_event
+                                        .send(AddEventToCurrentEvent::new_to_front(id.clone()));
                                 }
                             }
                         }