+++ src/game_events/mod.rs
@@ -3,7 +3,10 @@ use bevy::prelude::*;
 use serde::Deserialize;
 use std::collections::HashSet;
 
+mod add_event_to_current_event;
 mod event_display;
+mod event_effect_handler;
+mod event_shipwreck;
 mod event_travel;
 mod event_trigger_handler;
 mod event_visit_settlement;
@@ -20,6 +23,7 @@ pub struct GameEventAction {
 pub enum GameEventTriggerEventName {
     Travel,
     Settlement,
+    Shipwreck,
 }
 
 #[derive(Debug, Deserialize, PartialEq)]
@@ -32,6 +36,19 @@ pub struct GameEventTriggerCondition {
     pub chance: Option<f32>,
 }
 
+#[derive(Debug, Deserialize, Clone)]
+#[serde(deny_unknown_fields)]
+pub struct DamageEffect {
+    pub amount: u32,
+}
+
+#[derive(Deserialize, Debug, Clone)]
+#[serde(tag = "type", rename_all = "snake_case")]
+pub enum GameEventEffect {
+    DamageAnyShip(DamageEffect),
+    DamageAllShips(DamageEffect),
+}
+
 #[derive(Debug, Deserialize)]
 #[serde(deny_unknown_fields)]
 pub struct GameEvent {
@@ -39,13 +56,15 @@ pub struct GameEvent {
     pub title: String,
     pub image: String,
     pub trigger: Option<GameEventTriggerCondition>,
+    #[serde(default)]
+    pub effects: Vec<GameEventEffect>,
     pub text: String,
     pub actions: Vec<GameEventAction>,
 }
 
 #[derive(Default)]
 pub struct GameEventsState {
-    pub current_events: HashSet<String>,
+    pub current_events: Vec<String>,
     pub seen_events: HashSet<String>,
 }
 
@@ -55,18 +74,56 @@ pub struct TriggerEvent {
     pub scope: Option<String>,
 }
 
+#[derive(Debug)]
+pub struct TriggerEventEffect {
+    pub effect: GameEventEffect,
+}
+
+#[derive(Debug, Copy, Clone)]
+pub enum AddEventToCurrentEventPosition {
+    Front,
+    End,
+}
+
+#[derive(Debug)]
+pub struct AddEventToCurrentEvent {
+    pub id: String,
+    pub position: AddEventToCurrentEventPosition,
+}
+
+impl AddEventToCurrentEvent {
+    pub fn new(id: String) -> Self {
+        Self {
+            id,
+            position: AddEventToCurrentEventPosition::End,
+        }
+    }
+
+    pub fn new_to_front(id: String) -> Self {
+        Self {
+            id,
+            position: AddEventToCurrentEventPosition::Front,
+        }
+    }
+}
+
 pub struct GameEventsPlugin;
 
 impl Plugin for GameEventsPlugin {
     fn build(&self, app: &mut App) {
         app.init_resource::<GameEventsState>()
             .add_event::<TriggerEvent>()
+            .add_event::<TriggerEventEffect>()
+            .add_event::<AddEventToCurrentEvent>()
             .add_system_set(
                 SystemSet::on_update(RunningState::Paused)
                     .with_system(event_display::event_display),
             )
             .add_system(event_trigger_handler::event_trigger_handler)
+            .add_system(add_event_to_current_event::add_event_to_current_event)
+            .add_system(event_effect_handler::event_effect_handler)
             .add_system(event_travel::event_travel)
-            .add_system(event_visit_settlement::event_visit_settlement);
+            .add_system(event_visit_settlement::event_visit_settlement)
+            .add_system(event_shipwreck::event_shipwreck);
     }
 }