+++ src/types/mod.rs
@@ -1,5 +1,9 @@
 mod calculated_population_value;
+mod player;
 mod seasonal_amount;
+mod ship;
 
 pub use calculated_population_value::CalculatedPopulationValue;
+pub use player::Player;
 pub use seasonal_amount::SeasonalAmount;
+pub use ship::{Ship, ShipSize};